requirements.txt
django>=5
djangorestframework
djangorestframework-filters
django-cors-headers
python-decouple
pillow
drf-yasg
